Separate pointing and rogue pointing code
