+++ Content.Server/GameObjects/EntitySystems/RoguePointingSystem.cs
@@ -1,0 +1,26 @@
+﻿using Content.Server.GameObjects.Components.Pointing;
+using JetBrains.Annotations;
+using Robust.Shared.GameObjects;
+using Robust.Shared.GameObjects.Systems;
+
+namespace Content.Server.GameObjects.EntitySystems
+{
+    [UsedImplicitly]
+    public class RoguePointingSystem : EntitySystem
+    {
+        public override void Initialize()
+        {
+            base.Initialize();
+
+            EntityQuery = new TypeEntityQuery(typeof(RoguePointingArrowComponent));
+        }
+
+        public override void Update(float frameTime)
+        {
+            foreach (var entity in RelevantEntities)
+            {
+                entity.GetComponent<RoguePointingArrowComponent>().Update(frameTime);
+            }
+        }
+    }
+}
